--- test/isolated/set-commits-auto.test.ts
@@ -1,212 +1,0 @@
-/**
- * Isolated tests for setCommitsAuto
- *
- * Uses mock.module() for git helpers, so must run in isolation
- * to avoid polluting other test files' module state.
- */
-
-import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
-import { setAuthToken } from "../../src/lib/db/auth.js";
-import { setOrgRegion } from "../../src/lib/db/regions.js";
-import type { SentryRelease } from "../../src/types/index.js";
-import { mockFetch, useTestConfigDir } from "../helpers.js";
-
-useTestConfigDir("set-commits-auto-");
-
-mock.module("../../src/lib/git.js", () => ({
-  getRepositoryName: () => "getsentry/cli",
-  getHeadCommit: () => "abc123def456789012345678901234567890abcd",
-  isInsideGitWorkTree: () => true,
-  isShallowRepository: () => false,
-  getCommitLog: () => [],
-  getUncommittedFiles: () => [],
-  parseRemoteUrl: (url: string) => url,
-}));
-
-// Import after mock.module so the mocked git helpers are used
-const { setCommitsAuto } = await import("../../src/lib/api/releases.js");
-
-const SAMPLE_RELEASE: SentryRelease = {
-  id: 1,
-  version: "1.0.0",
-  shortVersion: "1.0.0",
-  status: "open",
-  dateCreated: "2025-01-01T00:00:00Z",
-  dateReleased: null,
-  firstEvent: null,
-  lastEvent: null,
-  ref: null,
-  url: null,
-  commitCount: 0,
-  deployCount: 0,
-  newGroups: 0,
-  authors: [],
-  projects: [
-    {
-      id: 1,
-      slug: "test-project",
-      name: "Test Project",
-      platform: "javascript",
-      platforms: ["javascript"],
-      hasHealthData: false,
-      newGroups: 0,
-    },
-  ],
-  data: {},
-  versionInfo: null,
-};
-
-const SAMPLE_REPO = {
-  id: "1",
-  name: "getsentry/cli",
-  url: "https://github.com/getsentry/cli",
-  provider: { id: "integrations:github", name: "GitHub" },
-  status: "active",
-};
-
-let originalFetch: typeof globalThis.fetch;
-
-beforeEach(async () => {
-  originalFetch = globalThis.fetch;
-  await setAuthToken("test-token");
-  setOrgRegion("test-org", "https://us.sentry.io");
-});
-
-afterEach(() => {
-  globalThis.fetch = originalFetch;
-});
-
-describe("setCommitsAuto", () => {
-  test("lists repos, discovers HEAD, fetches previous commit, and sends refs", async () => {
-    const withCommits = { ...SAMPLE_RELEASE, commitCount: 5 };
-    const requests: { method: string; url: string }[] = [];
-
-    globalThis.fetch = mockFetch(async (input, init) => {
-      const req = new Request(input!, init);
-      requests.push({ method: req.method, url: req.url });
-
-      // List org repositories (SDK uses /repos/ endpoint)
-      if (req.url.includes("/repos/")) {
-        expect(req.method).toBe("GET");
-        return new Response(JSON.stringify([SAMPLE_REPO]), {
-          status: 200,
-          headers: { "Content-Type": "application/json" },
-        });
-      }
-
-      // Previous release commit lookup
-      if (req.url.includes("/previous-with-commits/")) {
-        expect(req.method).toBe("GET");
-        return new Response(
-          JSON.stringify({
-            lastCommit: { id: "prev000000000000000000000000000000000000" },
-          }),
-          { status: 200, headers: { "Content-Type": "application/json" } }
-        );
-      }
-
-      // PUT refs on the release
-      expect(req.method).toBe("PUT");
-      expect(req.url).toContain("/releases/1.0.0/");
-      const body = (await req.json()) as {
-        refs: Array<{
-          repository: string;
-          commit: string;
-          previousCommit?: string;
-        }>;
-      };
-      expect(body.refs).toEqual([
-        {
-          repository: "getsentry/cli",
-          commit: "abc123def456789012345678901234567890abcd",
-          previousCommit: "prev000000000000000000000000000000000000",
-        },
-      ]);
-      return new Response(JSON.stringify(withCommits), {
-        status: 200,
-        headers: { "Content-Type": "application/json" },
-      });
-    });
-
-    const release = await setCommitsAuto("test-org", "1.0.0", "/tmp");
-
-    expect(release.commitCount).toBe(5);
-  });
-
-  test("throws ApiError when org has no repositories", async () => {
-    globalThis.fetch = mockFetch(
-      async () =>
-        new Response(JSON.stringify([]), {
-          status: 200,
-          headers: { "Content-Type": "application/json" },
-        })
-    );
-
-    await expect(setCommitsAuto("test-org", "1.0.0", "/tmp")).rejects.toThrow(
-      /No repository integrations/
-    );
-  });
-
-  test("throws ValidationError when no repo matches local remote", async () => {
-    const otherRepo = { ...SAMPLE_REPO, name: "getsentry/sentry" };
-    globalThis.fetch = mockFetch(
-      async () =>
-        new Response(JSON.stringify([otherRepo]), {
-          status: 200,
-          headers: { "Content-Type": "application/json" },
-        })
-    );
-
-    await expect(setCommitsAuto("test-org", "1.0.0", "/tmp")).rejects.toThrow(
-      /No Sentry repository matching/
-    );
-  });
-
-  test("paginates through multiple pages to find matching repo", async () => {
-    const withCommits = { ...SAMPLE_RELEASE, commitCount: 3 };
-    const otherRepo = { ...SAMPLE_REPO, id: "2", name: "getsentry/sentry" };
-    let repoRequestCount = 0;
-
-    globalThis.fetch = mockFetch(async (input, init) => {
-      const req = new Request(input!, init);
-
-      if (req.url.includes("/repos/")) {
-        repoRequestCount += 1;
-        if (repoRequestCount === 1) {
-          // First page: different repo, with a next cursor
-          return new Response(JSON.stringify([otherRepo]), {
-            status: 200,
-            headers: {
-              "Content-Type": "application/json",
-              Link: '<https://us.sentry.io/api/0/next/>; rel="next"; results="true"; cursor="page2:0:0"',
-            },
-          });
-        }
-        // Second page: the matching repo, no next cursor
-        return new Response(JSON.stringify([SAMPLE_REPO]), {
-          status: 200,
-          headers: { "Content-Type": "application/json" },
-        });
-      }
-
-      // Previous release commit lookup (no previous release)
-      if (req.url.includes("/previous-with-commits/")) {
-        return new Response(JSON.stringify({}), {
-          status: 200,
-          headers: { "Content-Type": "application/json" },
-        });
-      }
-
-      // PUT refs on the release
-      return new Response(JSON.stringify(withCommits), {
-        status: 200,
-        headers: { "Content-Type": "application/json" },
-      });
-    });
-
-    const release = await setCommitsAuto("test-org", "1.0.0", "/tmp");
-
-    expect(release.commitCount).toBe(3);
-    expect(repoRequestCount).toBe(2);
-  });
-});
