+++ test/lib/dsn/errors.test.ts
