+++ test/lib/dsn/resolver.test.ts
